Gardening: Better connection status gauge registration
Motivation
----------
Limiting the number of times a Micrometer MultiGauge is
registered improves the user experience for folks viewing
metrics in a JMX browser, since the gauge disappears
from the browser when it is reregistered.

Modifications
-------------
Only reregister the MultiGauge when the node list changes,
instead of when nodes connect or disconnect.

Change the connection status gauge to report a live value,
instead of the value from when the gauge was registered.

Change-Id: Icd982e036f985736acce32ad3c4e959772226913
Reviewed-on: http://review.couchbase.org/c/java-dcp-client/+/140161
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>