JDCP-196 Prevent "error dropped" log messages from Reactor
Motivation
----------
After upgrading Reactor from 3.3.10 to 3.4.3, a `doOnError` operator
is no longer sufficient to prevent Reactor from complaining that
an error was dropped.

Modifications
-------------
Use `onErrorResume` instead of `doOnError`. This prevents the error
from propagating through the chain and triggering the onErrorDropped
hook.

An alternative fix would have been to use the `subscribe` overload that
takes separate handlers for onNext, onError, onSuccess, etc, but that
doesn't look as clean, imo.

Change-Id: I8c574ec226db8e189fb843be6755a1106cedb524
Reviewed-on: http://review.couchbase.org/c/java-dcp-client/+/150889
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>