Gardening: Update integration tests to use Reactor 3.4.4
Motivation
----------
Align everything to use Reactor 3.4.4:
  * Upgrading Spring Boot from 2.3.4 to 2.4.4
  * Upgrade Couchbase Java SDK from 3.0.3 to 3.1.4.

Migrate the integration tests to JUnit 5 so they work with
Spring Boot 2.4.4.

Migrate the unit tests to JUnit 5 for consistency.

Use "openjdk:8-jdk-slim" instead of "openjdk:8-jdk-alpine"
to work around the JVM hard crash with netty native transports
on alpine with SDK 3.1.0 and later.

Change-Id: I4d01304adcc2a15df4df5f344cf5f37f30957078
Reviewed-on: http://review.couchbase.org/c/java-dcp-client/+/150893
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>