JDCP-223 Depend on core-io instead of repackaging Netty & Jackson
Motivation
----------
Adding core-io as a dependency is a prerequisite for using core-io for
smarter handling of DNS SRV.

As long as core-io is present, we might as well use its version
of Netty & Jackson instead of repackaging a duplicate version.

Modification
------------
The interesting changes are all on the build script.
Everything else is import changes and Javadoc fixes to pass the linter.

Followup changes will reduce duplicate code, and use core-io's DNS SRV
support to handle DNS SRV lookups (instead of requiring clients to resolve
DNS SRV before starting the DCP client like the must today).

Change-Id: Ie2d0a684762e9630da094c55a81edfba59e3e58c
Reviewed-on: https://review.couchbase.org/c/java-dcp-client/+/180328
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>