JDCP-217 Allow for configurations with only TLS ports
Motivation
----------
Support Couchbase clusters that advertise only TLS ports.

The DCP client is a good incubator for config parser changes that
might graduate to core-io later.

Modifications
-------------
Ground-up rewrite of the bucket config parser.

Refactor the config into ClusterConfig and CouchbaseBucketConfig
(which "has a" ClusterConfig).

Resolve alternate addresses and select TLS or non-TLS ports as the
config is parsed, with the help of NetworkSelector and PortSelector.

NodeInfo now has only TLS ports or only non-TLS ports, and only for
the relevant network. It no longer matters if the server's config
has only TLS ports, because the config parser ignores non-TLS ports
when TLS is enabled.

Change-Id: I095673f6501650121c1481e9ff2198d86e8aa994
Reviewed-on: https://review.couchbase.org/c/java-dcp-client/+/180801
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>