JDCP-224 When TLS is enabled, always use SASL mechanism "PLAIN"
Motivation
----------
* SCRAM-SHA provides no benefit when the channel is secured by TLS.
* PLAIN is faster, since it requires fewer round trips.
* PLAIN is the only mechanism that works for LDAP users.

Prior to this change, the DCP client did not work with LDAP users.
Now, LDAP users work if TLS is enabled.

Modifications
-------------
When TLS is enabled, skip the "list mechanisms" step, assume PLAIN
is supported, and use PLAIN to authenticate.

Disallow using PLAIN without TLS. This prevents a person-in-the-middle
from tricking the client into using PLAIN and sending the password
on an insecure channel.

Use the SCRAM SASL client from core-io. Remove the DCP client's
duplicate version.

Change-Id: I1bf843e5f6ac11b9324924f91f2e27a3165a3672
Reviewed-on: https://review.couchbase.org/c/java-dcp-client/+/181557
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>