Integration test: Don't sleep after creating bucket
Motivation
----------
Sleeping masks an interesting DCP edge case
related to inactive partitions.

When a partition is not yet active, nodes won't
return any results for getSeqnos(), and stream
requests fail with "not my vbucket."

In the integration tests, these things happen
because the bucket is not yet fully online.
We shouldn't just "fix" the tests by sleeping
after creating a bucket, because the same
issues with inactive partitions could happen
in a real-world scenario if the client connects
during a rebalance.

Change-Id: I24708d0acee93c37f31682c56022ab03d79be1fa
Reviewed-on: https://review.couchbase.org/c/java-dcp-client/+/203332
Reviewed-by: David Nault <david.nault@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>