JDCP-241 Client fails to resume stream after snapshot marker
Motivation
----------
If a channel receives a snapshot marker and then disconnects
(due to Couchbase node restart, for example), the client tries
to resume from an offset that is invalid because the seqno
is not within the snapshot boundaries.

When this happens, the client fails to resume streaming from
the affected partition.

Modifications
-------------
Track the pending snapshot separately, and only make it the
"current" snapshot after receiving the snapshot's first
data event.

Change-Id: I7c9cea750eb907611cd7b9ee4a4ccfb6d1b2e1f6
Reviewed-on: https://review.couchbase.org/c/java-dcp-client/+/203339
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>