JDCP-242 ConcurrentModificationException in LoggingEventConsumer
Motivation
----------
The LoggingEventConsumer used by MinimalEventBus
is not thread safe. In the Java/Scala/Kotlin SDKs that's
not a problem, since their EventBus always uses the same
thread to publish events.

The problem in is that the DCP client's custom MinimalEventBus
publishes events immediately on the caller's thread.
Multiple concurrent callers can lead to ConcurrentModification
inside LoggingEventConsumer due to concurrent HashMap updates.

Modifications
-------------
Add `synchronized` keyword to MinimalEventBus.publish().

Normally we'd look for an alternative to synchronizing,
but MinimalEventBus is used only during bootstrapping,
and not in any performance-sensitive code.

Change-Id: Ie4a2c251d571badb8d80021945247764ab114bbf
Reviewed-on: https://review.couchbase.org/c/java-dcp-client/+/205414
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>