JDCP-253 Log connection name at INFO level
Motivation
----------
Make it easier to correlate connections
in client and server logs.

Modifications
-------------
Add a channel attribute containing a description
of the channel (connection ID and remote address).

Make DcpPipeline responsible for assigning
connection name (instead of DcpConnectHandler)
so the channel description can be logged by any handler.

Wherever a handler does logging, prefix the log message
with the channel description (if available).

Change-Id: I25c130d51d6e14c966abc579b68d282d09e3e55f
Reviewed-on: https://review.couchbase.org/c/java-dcp-client/+/223837
Reviewed-by: Matt Ingenthron <ingenthr@gmail.com>
Tested-by: Build Bot <build@couchbase.com>